Remove TC4820 maximum power from config, return power as a float

+++ finesse/hardware/temperature/temperature_controller_base.py
@@ -60,7 +60,7 @@ class TemperatureControllerBase(DeviceBase):
 
     @property
     @abstractmethod
-    def power(self) -> int:
+    def power(self) -> float:
         """The current power output of the device."""
 
     @property
